# python-flask-server-generated/swagger_server/models/enumerate_offset.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class EnumerateOffset(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self):  # noqa: E501
        """EnumerateOffset - a model defined in Swagger

        """
        self.swagger_types = {
        }

        self.attribute_map = {
        }

    @classmethod
    def from_dict(cls, dikt) -> 'EnumerateOffset':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The EnumerateOffset of this EnumerateOffset.  # noqa: E501
        :rtype: EnumerateOffset
        """
        return util.deserialize_model(dikt, cls)